requirements.txt
Flask==2.2.5
gunicorn==20.1.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors